refactor: extract static/branding/PWA-asset routes into src/routes/static.ts
Part 1 of the src/index.ts route split (#661). Moves the no-auth,
no-validation, no-rate-limit routes — logo/OG/favicon/icon assets,
content-hashed CSS/JS, /health, /robots.txt, /sitemap.xml — into a new
Hono sub-router mounted at "/", matching the app.route() pattern already
used for authRoutes/dashboardRoutes/stripeWebhookRoutes. Behavior-preserving:
routes stay reachable through the exported `app`, so no test changes were
needed.
